+++ Assets/Scripts/Networking/RTSNetworkManager.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using Mirror;
+using UnityEngine;
+
+public class RTSNetworkManager : NetworkManager
+{
+    
+}
